# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MoneyQuest Personal Finance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  currency?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  xp: number;
  level: number;
  streak: number;
  isPremium: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserSettings {
  id: number;
  userId: number;
  language: string;
  currency: string;
  theme: string;
  /** @nullable */
  financialGoal?: string | null;
  notificationsEnabled: boolean;
  weeklyReportEnabled: boolean;
}

export interface UpdateUserSettingsInput {
  /** @nullable */
  language?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  theme?: string | null;
  /** @nullable */
  financialGoal?: string | null;
  /** @nullable */
  notificationsEnabled?: boolean | null;
  /** @nullable */
  weeklyReportEnabled?: boolean | null;
}

export interface UserStats {
  xp: number;
  level: number;
  streak: number;
  xpToNextLevel: number;
  totalChallengesCompleted: number;
  totalAchievements: number;
  financialHealthScore: number;
}

export interface Transaction {
  id: number;
  userId: number;
  type: string;
  amount: number;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  description: string;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  date: string;
  isRecurring: boolean;
  /** @nullable */
  recurringPeriod?: string | null;
  createdAt: string;
}

export interface CreateTransactionInput {
  type: string;
  amount: number;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  description: string;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  date: string;
  isRecurring?: boolean;
  /** @nullable */
  recurringPeriod?: string | null;
}

export interface UpdateTransactionInput {
  /** @nullable */
  type?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  subcategory?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /** @nullable */
  date?: string | null;
  /** @nullable */
  isRecurring?: boolean | null;
  /** @nullable */
  recurringPeriod?: string | null;
}

export interface Goal {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  icon?: string | null;
  targetAmount: number;
  currentAmount: number;
  /** @nullable */
  monthlyContribution?: number | null;
  /** @nullable */
  deadline?: string | null;
  status: string;
  progressPercentage: number;
  /** @nullable */
  projectedCompletionDate?: string | null;
  createdAt: string;
}

export interface CreateGoalInput {
  title: string;
  /** @nullable */
  icon?: string | null;
  targetAmount: number;
  currentAmount?: number;
  /** @nullable */
  monthlyContribution?: number | null;
  /** @nullable */
  deadline?: string | null;
}

export interface UpdateGoalInput {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  targetAmount?: number | null;
  /** @nullable */
  currentAmount?: number | null;
  /** @nullable */
  monthlyContribution?: number | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface ContributeGoalInput {
  amount: number;
}

export interface Budget {
  id: number;
  userId: number;
  category: string;
  monthlyLimit: number;
  currentSpent: number;
  remaining: number;
  percentUsed: number;
  status: string;
  month: number;
  year: number;
  createdAt: string;
}

export interface CreateBudgetInput {
  category: string;
  monthlyLimit: number;
  month?: number;
  year?: number;
}

export interface UpdateBudgetInput {
  /** @nullable */
  monthlyLimit?: number | null;
}

export interface Subscription {
  id: number;
  userId: number;
  name: string;
  amount: number;
  billingCycle: string;
  nextChargeDate: string;
  category: string;
  isEssential: boolean;
  annualCost: number;
  createdAt: string;
}

export interface CreateSubscriptionInput {
  name: string;
  amount: number;
  billingCycle: string;
  nextChargeDate: string;
  category: string;
  isEssential?: boolean;
}

export interface UpdateSubscriptionInput {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  billingCycle?: string | null;
  /** @nullable */
  nextChargeDate?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  isEssential?: boolean | null;
}

export interface Challenge {
  id: number;
  title: string;
  description: string;
  xpReward: number;
  type: string;
  difficulty: string;
  isCompleted: boolean;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface ChallengeResult {
  challenge: Challenge;
  xpEarned: number;
  /** @nullable */
  newLevel?: number | null;
  leveledUp: boolean;
}

export interface Achievement {
  id: number;
  title: string;
  description: string;
  icon: string;
  xpReward: number;
  isUnlocked: boolean;
  /** @nullable */
  unlockedAt?: string | null;
  category: string;
}

export interface DashboardInsights {
  totalBalance: number;
  incomeThisMonth: number;
  expensesThisMonth: number;
  savingsThisMonth: number;
  savingsRate: number;
  budgetHealthScore: number;
  recentTransactions: Transaction[];
  goalProgress: Goal[];
  upcomingSubscriptions: Subscription[];
  smartInsights: string[];
  streak: number;
  xp: number;
  level: number;
  dailyChallenge: Challenge;
}

export interface MonthlyData {
  month: string;
  income: number;
  expenses: number;
  savings: number;
}

export interface CategoryData {
  category: string;
  amount: number;
  percentage: number;
  transactionCount: number;
}

export interface SavingsCalculatorInput {
  initialAmount: number;
  monthlyContribution: number;
  annualRate: number;
  years: number;
  /** @nullable */
  inflationRate?: number | null;
  /** @nullable */
  yearlyIncrease?: number | null;
}

export interface YearlyProjection {
  year: number;
  balance: number;
  contributed: number;
  interest: number;
}

export interface SavingsProjection {
  futureValue: number;
  totalContributed: number;
  interestEarned: number;
  /** @nullable */
  inflationAdjustedValue?: number | null;
  yearlyData: YearlyProjection[];
}

export type GetTransactionsParams = {
  /**
   * @nullable
   */
  type?: string | null;
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  startDate?: string | null;
  /**
   * @nullable
   */
  endDate?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  offset?: number | null;
};

export type GetMonthlyInsightsParams = {
  /**
   * @nullable
   */
  months?: number | null;
};

export type GetCategoryInsightsParams = {
  /**
   * @nullable
   */
  startDate?: string | null;
  /**
   * @nullable
   */
  endDate?: string | null;
};