# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MoneyQuest Personal Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: Get all goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalInput"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalInput"
      responses:
        "200":
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /goals/{id}/contribute:
    post:
      operationId: contributeToGoal
      tags: [goals]
      summary: Add money to a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContributeGoalInput"
      responses:
        "200":
          description: Contribution added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /budgets:
    get:
      operationId: getBudgets
      tags: [budgets]
      summary: Get all budgets
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetInput"
      responses:
        "201":
          description: Budget created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetInput"
      responses:
        "200":
          description: Budget updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success
    RegisterInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        language:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
      required:
        - email
        - password
        - name
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatar:
          type: ["string", "null"]
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        isPremium:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - xp
        - level
        - streak
        - isPremium
        - createdAt
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        icon:
          type: ["string", "null"]
        targetAmount:
          type: number
        currentAmount:
          type: number
        monthlyContribution:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
        status:
          type: string
        progressPercentage:
          type: number
        projectedCompletionDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - title
        - targetAmount
        - currentAmount
        - status
        - progressPercentage
        - createdAt
    CreateGoalInput:
      type: object
      properties:
        title:
          type: string
        icon:
          type: ["string", "null"]
        targetAmount:
          type: number
        currentAmount:
          type: number
        monthlyContribution:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
      required:
        - title
        - targetAmount
    UpdateGoalInput:
      type: object
      properties:
        title:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        targetAmount:
          type: ["number", "null"]
        currentAmount:
          type: ["number", "null"]
        monthlyContribution:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
    ContributeGoalInput:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount
    Budget:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        category:
          type: string
        monthlyLimit:
          type: number
        currentSpent:
          type: number
        remaining:
          type: number
        percentUsed:
          type: number
        status:
          type: string
        month:
          type: integer
        year:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - category
        - monthlyLimit
        - currentSpent
        - remaining
        - percentUsed
        - status
        - month
        - year
        - createdAt
    CreateBudgetInput:
      type: object
      properties:
        category:
          type: string
        monthlyLimit:
          type: number
        month:
          type: integer
        year:
          type: integer
      required:
        - category
        - monthlyLimit
    UpdateBudgetInput:
      type: object
      properties:
        monthlyLimit:
          type: ["number", "null"]
...