# lib/api-zod/src/generated/types/updateGoalInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MoneyQuest Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateGoalInput {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  targetAmount?: number | null;
  /** @nullable */
  currentAmount?: number | null;
  /** @nullable */
  monthlyContribution?: number | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  status?: string | null;
}