# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MoneyQuest Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  xp: number;
  level: number;
  streak: number;
  isPremium: boolean;
  createdAt: string;
}